serverutil/keyserver: separate GCP parts from keyserver command
Following the same pattern we used for cmd/upspinserver and
serverutil/upspinserver, this change moves the core functionality of
cmd/keyserver into the new serverutil/keyserver package.
When we move the GCP-related packages to the gcp.upspin.io repository,
the cmd/keyserver command in the core repository will remain without
the cloud parts.

Reviewed-on: https://upspin-review.googlesource.com/9060
