+++ cmd/keyserver/main.go
@@ -4,33 +4,31 @@
 
 // Keyserver is a wrapper for a key implementation that presents it as an HTTP
 // interface.
-package main
+package main // import "upspin.io/cmd/keyserver"
 
 import (
 	"flag"
 	"net"
-	"net/http"
 
 	"upspin.io/cloud/gcpmetric"
-	"upspin.io/cloud/https"
 	cloudLog "upspin.io/cloud/log"
-	"upspin.io/config"
-	"upspin.io/errors"
 	"upspin.io/factotum"
 	"upspin.io/flags"
-	"upspin.io/key/inprocess"
-	"upspin.io/key/server"
 	"upspin.io/log"
 	"upspin.io/metric"
-	"upspin.io/rpc/keyserver"
+	"upspin.io/serverutil/keyserver"
 	"upspin.io/upspin"
 
 	// Load required transports
 	_ "upspin.io/key/transports"
+
+	// Possible storage backends.
+	_ "upspin.io/cloud/storage/disk"
+	_ "upspin.io/cloud/storage/gcs"
 )
 
 const (
-	// serverName is the upspin username for this server.
+	// serverName is the name of this server.
 	serverName = "keyserver"
 
 	// metricSampleSize is the size of the sample from which pick one metric
@@ -45,12 +43,10 @@ const (
 var (
 	testUser    = flag.String("test_user", "", "initialize a test `user` (localhost, inprocess only)")
 	testSecrets = flag.String("test_secrets", "", "initialize test user with the secrets in this `directory`")
-	// The format of the email config file must be lines: api key, incoming email provider user name and password.
-	mailConfigFile = flag.String("mail_config", "", "config file name for incoming email signups")
 )
 
 func main() {
-	flags.Parse(flags.Server, "kind", "project", "serverconfig")
+	flags.Register("project")
 
 	if flags.Project != "" {
 		cloudLog.Connect(flags.Project, serverName)
@@ -60,55 +56,11 @@ func main() {
 		svr, err := gcpmetric.NewSaver(flags.Project, metricSampleSize, metricMaxQPS, "serverName", serverName)
 		if err != nil {
 			log.Fatalf("Can't start a metric saver for GCP project %q: %s", flags.Project, err)
-		} else {
-			metric.RegisterSaver(svr)
 		}
+		metric.RegisterSaver(svr)
 	}
 
-	cfg, err := config.FromFile(flags.Config)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	// Create a new key implementation.
-	var key upspin.KeyServer
-	switch flags.ServerKind {
-	case "inprocess":
-		key = inprocess.New()
-	case "server":
-		key, err = server.New(flags.ServerConfig...)
-	default:
-		err = errors.Errorf("bad -kind %q", flags.ServerKind)
-
-	}
-	if err != nil {
-		log.Fatalf("Setting up KeyServer: %v", err)
-	}
-
-	// Special hack for bootstrapping the inprocess key server.
-	setupTestUser(key)
-
-	httpStore := keyserver.New(cfg, key, upspin.NetAddr(flags.NetAddr))
-	http.Handle("/api/Key/", httpStore)
-
-	if logger, ok := key.(server.Logger); ok {
-		http.Handle("/log", logHandler{logger: logger})
-	}
-	if *mailConfigFile != "" {
-		f := cfg.Factotum()
-		if f == nil {
-			log.Fatal("supplied config must include keys when -mail_config set")
-		}
-		h, err := newSignupHandler(f, key, *mailConfigFile, flags.Project)
-		if err != nil {
-			log.Fatal(err)
-		}
-		http.Handle("/signup", h)
-	} else {
-		log.Println("-mail_config not set, /signup deactivated")
-	}
-
-	https.ListenAndServeFromFlags(nil, "keyserver")
+	keyserver.Main(setupTestUser)
 }
 
 // isLocal returns true if the name only resolves to loopback addresses.
@@ -129,6 +81,8 @@ func isLocal(addr string) bool {
 	return true
 }
 
+// setupTestUser uses the -test_user and -test_secrets flags to bootstrap the
+// inprocess key server with an initial user.
 func setupTestUser(key upspin.KeyServer) {
 	if *testUser == "" {
 		return
